Fix running wgpu
