+++ Sources/AdaECS/System/Executors/SingleThreadedSystemsGraphExecutor.swift
@@ -45,7 +45,6 @@ public struct SingleThreadedSystemsGraphExecutor: SystemsGraphExecutor {
                 }
             }
 
-            print("execute node", currentNode.name)
             await executeSystem(
                 system: currentNode,
                 world: world,
